+++ app/Http/Controllers/PostsController.php
@@ -11,7 +11,7 @@ class PostsController extends Controller {
     public function store(Request $request) {
         $this->validate($request, [
             "board" => "required",
-            "title" => "required",
+            // "title" => "required",
             "content" => "required"
         ]);
         $board = Boards::where("slug", $request->board)->first();
@@ -19,14 +19,20 @@ class PostsController extends Controller {
         $post->board_id = $board->id;
         $post->title = $request->title;
         $post->name = $request->name ?? null;
-        $post->thread_id = $board->thread_id ?? null;
+        $post->thread_id = $request->thread_id ?? null;
+        $post->has_file = $request->has_file === 'true' ? true : false;
         $post->option = $request->option ?? null;
         $post->content = $request->content;
-        $post->has_file = true;
         $post->is_thread = $request->is_thread === 'true' ? true : false;
         $post->save();
 
-        if ($request->has("file")) {
+        if ($post->thread_id) {
+            $thread = Posts::find($request->thread_id);
+            $thread->replies += 1;
+            $thread->save();
+        }
+
+        if ($request->has("file") && $request->file->getSize()) {
             $upload    = new Uploads();
             $file      = $request->file;
             $fn        = $file->getClientOriginalName();
@@ -39,8 +45,18 @@ class PostsController extends Controller {
                 $upload->mime_type   = $mime;
                 $upload->save();
             }
+            $post->upload = $upload;
+        } else {
+            $post->upload = null;
         }
-        $post->upload = $upload;
         return $post;
+    }
+
+    public function grab_thread(Request $request) {
+        if (!$request->has("id")) {
+            return ["status" => "error", "message" => "Required parameter 'id' is missing", "data" => []];
+        }
+        $threads = Posts::with(["upload"])->where("id", $request->id)->orWhere("thread_id", $request->id)->orderBy("id", "asc")->get();
+        return ["status" => "ok", "data" => $threads];
     }
 }