+++ src/pages/test.svelte
@@ -1,9 +1,0 @@
-<script>
-    import { metatags } from "@sveltech/routify";
-    metatags.title = "My Routify app";
-    metatags.description = "Description coming soon...";
-</script>
-
-<main>
-    <h1>Oi what's up cunt</h1>
-</main>